Safari-609.1.11.1/LayoutTests/fast/dom/Window/window-postmessage-clone-really-deep-array-expected.txt

Tests that we abort cloning overdeep arrays.

FAIL: 'postMessage(reallyDeepArray)' threw RangeError: Maximum call stack size exceeded., was expecting 25
PASS: eventData is done of type string
